Fix cpp tests and speculos tests
cpp_test: Move functions that should not be seen by cpp tests outside
hdpath.c. Set show_address variable. Add last warning to expected output.
Speculos: Delete slot at the end of test-slot-transactio-interaction
