+++ app/src/apdu_handler.c
@@ -94,7 +94,9 @@ __Z_INLINE void handleSign(volatile uint32_t *flags, volatile uint32_t *tx, uint
         THROW(APDU_CODE_DATA_INVALID);
     }
 
-    loadAddressCompareHdPathFromSlot();
+    show_address = show_address_none;
+    loadAddressCompareHdPathFromSlot();    
+
     //if we found matching hdPath on slot 0
     if (show_address == show_address_yes) {
         checkAddressUsedInTx();
